undo_move works and some simple tests from kingmoves
added som tests for king moves. on of them commented out for future improvments. fixing that testcase would likely speed up the move generation.
Furthermore added make_move and undo_last_move with a test to count the number of possible moves from a position. currently it fails this test on the forth halfmove.
Yet to is to implement special moves like castle, catch sudo-leagal moves and checkmate.