Trasposition Table added to Search
Added a Transposition table to the search. it reduces the amount of searched positions. However in some small tests it even tho ir reduces the amount or seached positions it increeces runtime by between 2 to 5%.
This is likely due to the tested positions was the entire game and the table will preform much better towards the endgame. Another factor is, currently it directly searches to depth 4 and not deeper.
A part of the tables function is to save the best move in each state. This will speed up the runtime when iterative deepening is implemented.

A bug where zobrist value didnt contain which turn it was. this lead to false hits in the Transposition table