Iterative deepening and quiet search
Implemented both iterative deepeing and quiet search. Currently the searchtime is set to 200 milliseconds. Without a limit in search time the quiet search for depth 4 took up to a minute.
That is why iterative deepining and searchtime limit was needed fo the bot to behave in a predicable way

Bugfix: Undo move didnt update the capture map of the opponent wich resulted in castle being possible in some situations which it wasnt supposed to be.