rook-files, doubled_pawns and mobility, Checkmate/stalemate
Implemented mobility score, score for past pawns and rooks on open files. laid the goundwork for adding open bishop files

Searcher can now see the diffrence between checkmate and stalemate. It can also see the if it repeats a position that have priviusly accured in the game to avoid three fold repetition. It still accepts stalemate and tree fold if it thinks it is in a inferior position.

Uci Position command now accepts both "fen" and no "fen" string when parsing a command without startpos flagg. The previous implementation didnt expect the fen flagg and therefore could not parse position command from fen