Uci info now supported
The program is now multithreaded with one thread for input and one for output. This allows the bot go send info messages during a seach. It reveals the current best move and the current best line and the evaluation in centipawns.

The go command now supports movetime and searchmoves. this allows for better debuging and a more accurate test by STS_rating