Faster movegeneration
Main change is faster move generation. The new move gen is between 5 to 10 times faster depending on the branching factor. the old depended on the branchingfactor but the new can get leagal moves from only the current position. This is done by calculating pinns to avoid descoverd check on your self and nbr of checks to see if only the king is allowed to move.

A few more tests added which the new movegeneration missed.

Bugfix:
Misstaken mate when the apponent only had one move. the move was filtered out before the check if the opponent doesnt have any leagal moves.

Draw by repetition is now avoided. previusly the search looked up the position in the transposition table before checking the history. Along with history not being added correcctly this resulted in failure to see three-fold-repetition.