Opening book, nullmove proning, check extentions
Opening book added. Not the optimal way of adding the book but it does the job. The opening book should be added as a polygot to the gui but now the program includes a book too.

Nullmove proning added. Doesnt add mutch elo tho together with check extentions it beats the previous version 53 win 14 draw 33 loss. With some small amount of testig this adds quite a bit of unstability. For now this is not a problem but in some positions it can return 3 or 4 diffrent moves as its best move. This is something to look in to in the future. Additionaly, something to see zugzwang need to be added to not fall into traps. Currently nullmove proning is only used when there is more than 10 pices on the board. This is to counteract zugzwang which most often occure in the endgame.

Check extentions. Doesnt add any notiable elo but together with Nullmove proning it appears to be an improvement. Currently max extentions is set to 2 but this needs to be tested further.