Magic bitboards
Implemented magic bitboards in movegeneration. The improvement is higly disappointing with only around a 4% improvement.
The way the tables are set up is not very good and the tests cant be run on multiple threads. The problem is that the const_eval_limit is reached and it cant be calculated during compiletime. TODO look into building the tables with buildscripts.